+++ src/pykeen/models/unimodal/mure.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
 
 """Implementation of MuRE."""
+
 from typing import Any, ClassVar, Mapping, Optional
 
 from torch.nn.init import normal_, uniform_, zeros_
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn import EmbeddingSpecification
+from ...nn.emb import EmbeddingSpecification
 from ...nn.modules import MuREInteraction
 from ...typing import Hint, Initializer
