+++ src/pykeen/models/unimodal/pair_re.py
@@ -9,7 +9,7 @@ from torch.nn.init import uniform_
 
 from ..nbase import ERModel
 from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
-from ...nn import EmbeddingSpecification
+from ...nn.emb import EmbeddingSpecification
 from ...nn.modules import PairREInteraction
 from ...typing import Hint, Initializer, Normalizer
