+++ src/pykeen/models/unimodal/rotate.py
@@ -9,7 +9,7 @@ import torch.autograd
 
 from ..base import EntityRelationEmbeddingModel
 from ...losses import Loss
-from ...nn import EmbeddingSpecification
+from ...nn.emb import EmbeddingSpecification
 from ...nn.init import init_phases, xavier_uniform_
 from ...nn.norm import complex_normalize
 from ...regularizers import Regularizer
