+++ src/pykeen/nn/__init__.py
@@ -2,12 +2,9 @@
 
 """PyKEEN internal "nn" module."""
 
-from . import init
-from .emb import Embedding, EmbeddingSpecification, RepresentationModule
+from . import init, norm
 
 __all__ = [
-    'Embedding',
-    'EmbeddingSpecification',
-    'RepresentationModule',
     'init',
+    'norm',
 ]
