+++ src/pykeen/nn/message_passing.py
@@ -13,6 +13,8 @@ from torch.nn import functional
 
 __all__ = [
     "Decomposition",
+    'BasesDecomposition',
+    'BlockDecomposition',
     "decomposition_resolver",
 ]
